pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "leaguesphere"
version = "2.17.3-rc.88"
requires-python = ">=3.12"
dependencies = [
    "Django==6.0.2",
    "django-crispy-forms==2.5",
    "crispy-bootstrap5==2025.6",
    "numpy==2.4.2",
    "pandas==3.0.1",
    "django-pandas==0.6.7",
    "djangorestframework==3.16.1",
    "django-autocomplete-light==3.12.1",
    "django-formtools==2.5.1",
    "Markdown==3.10.2",
    "django-cors-headers==4.9.0",
    "Pillow==12.1.1",
    "django-rest-knox==5.0.2",
    "python-dotenv==1.2.1",
    "mysqlclient==2.2.8",
    "requests==2.32.5",
    "django-health-check==4.0.4",
]

[project.optional-dependencies]
test = [
    "django-webtest==1.9.14",
    "pytest==9.0.2",
    "pytest-django==4.12.0",
    "pytest-cov==7.0.0",
    "factory-boy==3.3.3",
    "django-debug-toolbar==6.2.0",
    "WebTest==3.0.7",
    "bump2version==1.0.1",
    "black==26.1.0",
    "types-factory_boy==0.4.1",
]
prod = [
    "gunicorn>=22.0.0",
]

[tool.setuptools.packages.find]
include = [
    "accounts*",
    "gamedays*",
    "league_manager*",
    "league_table*",
    "liveticker*",
    "locust_testing*",
    "officials*",
    "passcheck*",
    "scorecard*",
    "teammanager*",
    "fe_template*"
]

[tool.bumpversion]
current_version = "2.17.3-rc.88"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)(\\-rc\\.(?P<rc_build>\\d+))?"
serialize = [
    "{major}.{minor}.{patch}-rc.{rc_build}",
    "{major}.{minor}.{patch}"
]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
ignore_missing_files = false
tag = true
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"
moveable_tags = []
commit_args = ""
setup_hooks = []
pre_commit_hooks = ["uv lock", "git add uv.lock"]
post_commit_hooks = []

[tool.bumpversion.parts.release]
optional_value = "stable"
first_value = "rc"
values = ["rc", "stable"]

[tool.bumpversion.parts.rc_build]
first_value = "1"

[[tool.bumpversion.files]]
filename = "league_manager/__init__.py"

[[tool.bumpversion.files]]
filename = "liveticker/package.json"

[[tool.bumpversion.files]]
filename = "passcheck/package.json"

[[tool.bumpversion.files]]
filename = "scorecard/package.json"

[[tool.bumpversion.files]]
filename = "gameday_designer/package.json"

[[tool.bumpversion.files]]
filename = "dashboard/package.json"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'
